Fix memory-safety bugs in core parsing and POSIX clock (#3538)
Code review of the library surfaced five latent defects (none duplicating
existing .github/triage items). Each is fixed and validated against the
core/IoT unit test suites (all passing).

* az_json_token.c (az_json_string_unescape): the destination bounds guard
  used `position > destination_size`, allowing a one-byte out-of-bounds
  write at index == destination_size before returning an over-long span.
  Changed to `>=` to match the canonical check used elsewhere (CWE-787).

* az_http_response.c (az_http_response_get_next_header): the header-value
  scan loop read reader[offset] inside a `while (true)` with no bound on
  offset, unlike the sibling field-name/OWS loops. A malformed or truncated
  response whose final header value lacks a terminating CR read past the end
  of the buffer. Added an `offset >= reader_size` guard (CWE-125).

* az_http_response.c (_az_get_http_status_line): a status line whose status
  code is immediately followed by LF made offset == 0, so the reason-phrase
  slice requested end index -1 (negative-size span). Guarded with
  `offset > 0 ? offset - 1 : 0`.

* az_posix.c (az_platform_clock_msec): `ts.tv_sec * 1000` was evaluated in
  time_t width (the constants are int enum values) and only cast to int64_t
  afterward. On 32-bit time_t targets this overflows once monotonic uptime
  exceeds ~24.8 days, corrupting retry backoff, SAS expiry, and context
  timeouts. Cast tv_sec/tv_nsec to int64_t before the arithmetic (CWE-190).

* az_http_request.c (az_http_request_init): url_length was never validated
  against the url span size, so the query scan (and later az_span_slice
  calls) could read out of bounds on caller misuse. Added
  `_az_PRECONDITION_RANGE(0, url_length, az_span_size(url))`, matching the
  convention of sibling init functions.

+++ sdk/src/azure/platform/az_posix.c
@@ -21,8 +21,8 @@ AZ_NODISCARD az_result az_platform_clock_msec(int64_t* out_clock_msec)
     return AZ_ERROR_CLOCK;
   }
 
-  *out_clock_msec = (int64_t)((ts.tv_sec * _az_TIME_MILLISECONDS_PER_SECOND)
-                              + (ts.tv_nsec / _az_TIME_NANOSECONDS_PER_MILLISECOND));
+  *out_clock_msec = ((int64_t)ts.tv_sec * _az_TIME_MILLISECONDS_PER_SECOND)
+      + ((int64_t)ts.tv_nsec / _az_TIME_NANOSECONDS_PER_MILLISECOND);
 
   return AZ_OK;
 }
